Bug fixes for compiling on GPU
