+++ dynet/gpu-nodes-affinetransform.cu
@@ -1,3 +1,3 @@
 // This is a dummy file that contains the same content as nodes.cc but compiled
 // on CUDA
-#include "nodes-affinetranform.cc"
+#include "nodes-affinetransform.cc"
