+++ dynet/nodes.cc
@@ -163,34 +163,6 @@ size_t MinDimension::aux_storage_size() const {
 
 // ===== Functions to be compiled on both CPU and GPU
 
-#ifdef __CUDACC__
-inline void CUDAMatrixMultiply(const Device_GPU & dev, const Tensor& l, const Tensor& r, Tensor& y, const float* acc_scalar) {
-  if(l.d.bd == 1 && r.d.bd == y.d.bd) {
-    // If the left side has one batch, multiply by columns
-    // [x, z, b] = [x, y] * [y, z, b]
-    // -> [x, z*b] = [x, y], [y, z*b]
-    CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_N, CUBLAS_OP_N,
-          y.d.rows(), y.d.cols() * y.d.batch_elems(), l.d.cols(),
-          kSCALAR_ONE,
-          l.v, l.d.rows(),
-          r.v, r.d.rows(),
-          acc_scalar, y.v, y.d.rows()));
-  } else {
-    // Otherwise, loop over the batches
-    DYNET_ASSERT(r.d.bd != 1 || r.d.bd != l.d.bd,
-                 "Number of batch elements in matrix multiply must match, but got: " << r.d.bd << ", " << l.d.bd);
-    for(unsigned b = 0; b < y.d.bd; ++b) {
-      CUBLAS_CHECK(cublasSgemm(dev.cublas_handle, CUBLAS_OP_N, CUBLAS_OP_N,
-            y.d.rows(), y.d.cols(), l.d.cols(),
-            kSCALAR_ONE,
-            l.batch_ptr(b), l.d.rows(),
-            r.batch_ptr(b), r.d.rows(),
-            acc_scalar, y.batch_ptr(b), y.d.rows()));
-    }
-  }
-}
-#endif
-
 template<class MyDevice>
 void AddVectorToAllColumns::forward_dev_impl(const MyDevice & dev, const vector<const Tensor*>& xs, Tensor& fx) const {
   // Broadcasting is slow on CPU, so split codepaths
